type hint N_eff from int to float (techincally should be int, but floats work too for the simulation)

+++ Projects/Project 2/bbn/background.py
@@ -7,7 +7,7 @@ class Background:
     """Class for the background functions and variables"""
 
     def __init__(
-        self, unit: str = "CGS", N_eff: int = 3, Omega_b0: float = 0.05
+        self, unit: str = "CGS", N_eff: float = 3.0, Omega_b0: float = 0.05
     ) -> None:
         """Initializes the class with the given parameters.
